Rework option filtering
So far options were filtered by Angular's default filters. This caused a
lot of problems, as the angular rendering ran asynchronously with the
rest of the updates happening in our code, which in many ways depended
on the filtering produced by the rendering.

This moves the filtering logic to the clientside proper, making things
nice and serial.